+++ artifacts/server/marketplace/src/main/java/com/worksonourmachines/marketplace/tutorprofile/service/MarketplaceTutorProfileService.java
@@ -67,8 +67,8 @@ public class MarketplaceTutorProfileService {
             @Nullable String topicId,
             @Nullable List<String> languages,
             @Nullable List<SharedMarketplaceLocation> locations,
-            @Nullable Float minRate,
-            @Nullable Float maxRate,
+            @Nullable Integer minRate,
+            @Nullable Integer maxRate,
             @Nullable Float minRating,
             @Nullable List<SharedMarketplaceWeekday> weekdays,
             @Nullable SharedMarketplaceTutorSort sort) {
@@ -128,8 +128,8 @@ public class MarketplaceTutorProfileService {
             @Nullable String topicId,
             @Nullable List<String> languages,
             @Nullable List<SharedMarketplaceLocation> locations,
-            @Nullable Float minRate,
-            @Nullable Float maxRate,
+            @Nullable Integer minRate,
+            @Nullable Integer maxRate,
             @Nullable Float minRating,
             @Nullable List<SharedMarketplaceWeekday> weekdays) {
         return (root, query, criteriaBuilder) -> {
